Revert "Separate classes and interfaces"
This reverts commit 185dc5db46c943c0be64de2baf9a981034ad5ccb.
