--- SolidPractice/AreaEnglishMessenger.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolidPractice
-{
-    public class AreaEnglishMessenger : IMessenger
-    {
-        public void Message(double area)
-        {
-            Console.WriteLine("The total area is " + area);
-        }
-    }
-}
