--- SolidPractice/IMessenger.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolidPractice
-{
-    public interface IMessenger
-    {
-        void Message(double val);
-
-    }
-}
