--- SolidPractice/IShape.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolidPractice
-{
-    public interface IShape
-    {
-        int Width { get; set; }
-    }
-}
