+++ SolidPractice/SolidPractice.csproj
@@ -43,23 +43,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AreaCalculator.cs" />
-    <Compile Include="AreaEnglishMessenger.cs" />
-    <Compile Include="Circle.cs" />
-    <Compile Include="Cube.cs" />
-    <Compile Include="Equilateral_Triangle.cs" />
-    <Compile Include="IArea.cs" />
-    <Compile Include="IAreaCalculator.cs" />
-    <Compile Include="IMessenger.cs" />
-    <Compile Include="IShape.cs" />
-    <Compile Include="IVolume.cs" />
-    <Compile Include="IVolumeCalculator.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Square.cs" />
-    <Compile Include="VolEnglishMessenger.cs" />
-    <Compile Include="VolFrenchMessenger.cs" />
-    <Compile Include="VolumeCalculator.cs" />
+    <Compile Include="SolidClass.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
