--- SolidPractice/VolFrenchMessenger.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolidPractice
-{
-    public class VolFrenchMessenger : IMessenger
-    {
-        public void Message(double volume)
-        {
-            Console.WriteLine("Le volume total est " + volume);
-        }
-    }
-}
