--- SolidPractice/VolumeCalculator.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolidPractice
-{
-    public class VolumeCalculator : IVolumeCalculator
-    {
-        public double CalculateVolume(IList<IVolume> shapes)
-        {
-            double totalVolume = 0;
-            foreach (var shape in shapes)
-            {
-                totalVolume += shape.GetVolume();
-            }
-            return totalVolume;
-        }
-    }
-}
